Move test_legvander to separate files
