+++ projection_matrix_gpu.py
@@ -24,3 +24,25 @@ def hermevander_wrapper(x, deg):
     hermevander[numblocks, blocksize](x, deg, output)
     return cp.squeeze(output)
 
+@cuda.jit
+def legvander(x, deg, output_matrix):
+    i = cuda.grid(1)
+    stride = cuda.gridsize(1)
+    for i in range(i, x.shape[0], stride):
+        output_matrix[i][0] = 1
+        output_matrix[i][1] = x[i]
+        for j in range(2, deg + 1):
+            output_matrix[i][j] = (output_matrix[i][j-1]*x[i]*(2*j - 1) - output_matrix[i][j-2]*(j - 1)) / j
+
+def legvander_wrapper(x, deg):
+    """Temporary wrapper that allocates memory and defines grid before calling legvander.
+    Probably won't be needed once cupy has the correpsponding legvander function.
+
+    Input: Same as cpu version of legvander
+    Output: legvander matrix, cp.ndarray
+    """
+    output = cp.ndarray((len(x), deg + 1))
+    blocksize = 256
+    numblocks = (len(x) + blocksize - 1) // blocksize
+    legvander[numblocks, blocksize](x, deg, output)
+    return output
